Added USE=+pixbuf-loader to libsrvg
Previously, this was not configurable, and a dependency on gdk-pixbuf
was hard-coded into the ebuild. However, the dependency should have
specified gdk-pixbuf[X], which it did not. This change takes care of
that problem as well.

Would be nice to say "+" on pixbuf-loader only if gtk+ is requested.
Wonder if there's a way to do that?